minor refactoring for more portability, added 'logic.Configuration' package to house constants and paths
