+++ src/logic/Controllers/GraphicsController.java
@@ -1,7 +1,7 @@
 package logic.Controllers;
 
 import javafx.scene.canvas.GraphicsContext;
-import logic.Model.Constants;
+import logic.Configuration.Constants;
 
 import java.util.concurrent.TimeUnit;
