--- streakergame.properties
@@ -1,4 +1,0 @@
-path.variable.kotlin_bundled=/Applications/IntelliJ IDEA.app/Contents/plugins/Kotlin/kotlinc
-path.variable.maven_repository=/Users/kylerramsey/.m2/repository
-jdk.home.9=/Library/Java/JavaVirtualMachines/jdk-9.jdk/Contents/Home
-javac2.instrumentation.includeJavaRuntime=false
